add manual json marshaller for AccessTokenRequest
