+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/client/TokenRequest.kt
@@ -2,7 +2,7 @@ package org.http4k.security.oauth.client
 
 import org.http4k.core.Body
 import org.http4k.core.Uri
-import org.http4k.format.Moshi.auto
+import org.http4k.format.OAuthMoshi.auto
 import org.http4k.security.oauth.core.RefreshToken
 import org.http4k.security.oauth.server.ClientId
