+++ http4k-security/oauth/src/main/kotlin/org/http4k/security/oauth/server/AccessTokenResponseRenderer.kt
@@ -4,9 +4,9 @@ import org.http4k.core.Body
 import org.http4k.core.Response
 import org.http4k.core.Status
 import org.http4k.core.with
+import org.http4k.format.OAuthMoshi.auto
 import org.http4k.security.AccessTokenDetails
 import org.http4k.security.AccessTokenResponse
-import org.http4k.security.oauth.server.OAuthServerMoshi.auto
 import org.http4k.security.openid.IdToken
 
 fun interface AccessTokenResponseRenderer : (AccessTokenDetails) -> Response
